JavaScript - Qualtrics Site Intercept - Canvas
Download this file and replace the three instances of '*************' on line 82 with your Site Intercept ID. The Site Intercept ID is found at the bottom of the Deployment code for your Site Intercept in Qualtrics. It is the string of text that comes directly after ZN_. Also on line 82, replace '[brandID]' with your Qualtrics Brand Id.